% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjbart.R
\name{AMCE}
\alias{AMCE}
\title{Average Marginal Component Effect Estimation with credible Interval}
\usage{
AMCE(
  data,
  model,
  attribs,
  ref_levels,
  method = "bayes",
  alpha = 0.05,
  cores = 1,
  skip_checks = FALSE
)
}
\arguments{
\item{data}{A data.frame, containing all attributes, covariates, the outcome and id variables to analyze.}

\item{model}{A model object, the result of running \code{cjbart()}}

\item{attribs}{Vector of attribute names for which IMCEs will be predicted}

\item{ref_levels}{Vector of reference levels, used to calculate marginal effects}

\item{method}{Character string, setting the variance estimation method to use. When method is "parametric", a typical combined variance estimate is employed; when \code{method = "bayes"}, the 95\% posterior interval is calculated; and when \code{method = "rubin"}, combination rules are used to combine the variance analogous to in multiple imputation analysis.}

\item{alpha}{Number between 0 and 1 -- the significance level used to compute confidence/posterior intervals. When \code{method = "bayes"}, the posterior interval is calculated by taking the alpha/2 and (1-alpha/2) quantiles of the posterior draws. When \code{method = "rubin"}, the confidence interval equals the IMCE +/- \code{qnorm(alpha/2)}. By default, alpha is 0.05 i.e. generating a 95\% confidence/posterior interval.}

\item{cores}{Number of CPU cores used during prediction phase}

\item{skip_checks}{Boolean, indicating whether to check the structure of the data (default = \code{FALSE}). Only set this to \code{TRUE} if you are confident that the data is structured appropriately}
}
\value{
\code{AMCE} returns an object of type "cjbart", a list object.
\item{amces}{A data.frame containing the average marginal component effects}
\item{alpha}{The significance level used to compute the credible interval}
}
\description{
\code{AMCE} calculates the average marginal component effects from a BART-estimated conjoint model.
}
\details{
The AMCE estimates are the average of all computed OMCEs.
}
\examples{
subjects <- 5
rounds <- 2
profiles <- 2
obs <- subjects*rounds*profiles

fake_data <- data.frame(A = sample(c("a1","a2"), obs, replace = TRUE),
                        B = sample(c("b1","b2"), obs, replace = TRUE),
                        id1 = rep(1:subjects, each=rounds),
                        stringsAsFactors = TRUE)

fake_data$Y <- sample(c(0,1), obs, replace = TRUE)

cj_model <- cjbart(data = fake_data,
                   Y = "Y",
                   id = "id1")

## Skip if not Unix due to longer CPU time
if (.Platform$OS.type=='unix') {

  het_effects <- IMCE(data = fake_data,
                      model = cj_model,
                      attribs = c("A","B"),
                      ref_levels = c("a1","b1"),
                      cores = 1)

  summary(het_effects)
}


}
\seealso{
\code{\link[=cjbart]{cjbart()}}
}
