% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjbart.R
\name{cjbart}
\alias{cjbart}
\title{Generate Conjoint Model Using BART}
\usage{
cjbart(data, Y, id = NULL, ...)
}
\arguments{
\item{data}{A data.frame or coercible, containing all attributes, covariates, the outcome and id variables to analyse.}

\item{Y}{Character string -- the outcome variable}

\item{id}{Character string -- the variable identifying individual respondents}

\item{...}{Other arguments passed to pbart}
}
\value{
Tibble summarising the average marginal component effect, the minimum and maximum values, and standard deviations for each attribute-level.
}
\description{
A wrapper for the \code{BART::pbart()} function.
}
\details{
Please note, \code{cjbart} currently only works for a binary outcome.
}
\seealso{
\code{\link[BART:pbart]{BART::pbart()}}
}
