% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimp.R
\name{het_vimp}
\alias{het_vimp}
\title{Estimate Variable Importance Metrics for \code{cjbart} Object}
\usage{
het_vimp(model, outcomes = NULL, covars = NULL)
}
\arguments{
\item{model}{Object of class \code{cjbart}, the result of running \code{\link[=IMCE]{IMCE()}}}

\item{outcomes}{An optional vector of attribute levels to generate importance metrics for. By default, all attribute-levels are analyzed.}

\item{covars}{An optional vector of covariates to include in the importance metric check. By default, all covariates are included in each importance model.}
}
\value{
A long data.frame of variable importance scores for each covariate per attribute-level, as well as the estimated 95\% confidence intervals for each score.
}
\description{
Estimates random forest variable importance scores for multiple attribute-levels of a conjoint experiment.
}
\details{
Having generated a schedule of individual-level marginal component effect estimates, for each attribute-level this function fits a random forest model using the supplied covariates as predictors. It then calculates a variable importance measure (VIMP) for each covariate. The VIMP assesses how important each covariate is in terms of partitioning the predicted individual-level effects distribution, and can thus be used as an indicator of where heterogeneity exists.
To recover our VIMP measure, we used permutation-based importance metrics recovered from random forest models (see \code{\link[randomForestSRC:rfsrc]{randomForestSRC::rfsrc()}}. To permute the data, this function uses random node assignment where cases are randomly assigned to a daughter node whenever a tree splits on the target variable (see \insertCite{ishwaran2008random}{cjbart}). Importance is defined in terms of how this random assignment degrades the performance of the forest. Higher degradation indicates a variable is more important to the prediction.
Variance estimates for the importance measures are recovered using the delete-d jackknife estimator developed by \insertCite{ishwaran2019standard;textual}{cjbart} and implemented in \pkg{randomForestSRC} \insertCite{randomforestsrc}{cjbart}. The jackknife method has inherent bias correction properties, making it particularly effective for variable selection exercises such as in identifying drivers of heterogeneity.
}
