% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjbart.R
\name{IMCE}
\alias{IMCE}
\title{Heterogeneous Effects Analysis of Conjoint Results}
\usage{
IMCE(
  data,
  model,
  attribs,
  ref_levels,
  Y,
  id,
  method = "bayes",
  alpha = 0.05,
  keep_omce = FALSE,
  cores = 1
)
}
\arguments{
\item{data}{A data.frame or coercible, containing all attributes, covariates, the outcome and id variables to analyse.}

\item{model}{A model object, the result of running \code{cjbart}}

\item{attribs}{Vector of attribute names}

\item{ref_levels}{Vector of reference levels, used to calculate marginal effects}

\item{Y}{Character string -- the outcome variable}

\item{id}{Character string -- the variable identifying individual respondents}

\item{method}{Character string -- the variance estimation method to use. When method = "parametric", a typical combined variance estimate is employed; when method = "bayes", the 95th posterior interval is calculated; and when method = "rubin", combination rules are used to combine the variance analogous to in multiple imputation analysis.}

\item{alpha}{Number between 0 and 1 -- the significance level used to compute confidence/posterior intervals. When method = "bayes", the posterior interval is calculated by taking the alpha/2 and (1-alpha/2) quantiles of the posterior draws. When method = "rubin", the confidence interval equals the IMCE +/- qnorm(alpha/2). By default, alpha = 0.05 i.e. generating a 95\% confidence/posterior interval.}

\item{keep_omce}{Boolean, indicating whether to keep the OMCE-level results (default = \code{FALSE})}

\item{cores}{Number of CPU cores used during prediction phase}
}
\value{
\code{OMCE} returns an object of type "cjbart", a list object.
\item{omce}{A data.frame containing the observation-level marginal effects}
\item{imce}{A data.frame containing the individual-level marginal effects}
\item{att_levels}{A vector containing the attribute levels}
}
\description{
\code{IMCE} calculates the individual-level marginal component effects from a BART-estimated conjoint model.
}
\details{
The OMCE estimates are the result of subtracting the predicted value of each observation under the reference-level category from the predicted value of each observation under the given attribute level.
If an attribute has \code{k} levels, then this will yield \code{k-1} estimates per observation.
The IMCE is the average of the OMCEs for each individual within the data.
}
\seealso{
\code{\link[=cjbart]{cjbart()}}
}
