% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjbart.R
\name{OMCE}
\alias{OMCE}
\title{Heterogeneous Effects Analysis of Conjoint Results}
\usage{
OMCE(data, model, attribs, ref_levels, Y_var, id_var, cores = 1)
}
\arguments{
\item{data}{A data.frame or coercible, containing all attributes, covariates, the outcome and id variables to analyse.}

\item{model}{A model object, the result of running \code{cjbart}}

\item{attribs}{Vector of attribute names}

\item{ref_levels}{Vector of reference levels, used to calculate marginal effects}

\item{Y_var}{Character string -- the outcome variable}

\item{id_var}{Character string -- the variable identifying individual respondents}

\item{cores}{Number of CPU cores used during prediction phase}
}
\value{
\code{OMCE} returns an object of type "cjbart", a list object.
\item{omce}{A data.frame containing the observation-level marginal effects}
\item{imce}{A data.frame containing the individual-level marginal effects}
\item{att_levels}{A vector containing the attribute levels}
}
\description{
\code{OMCE} calculates the marginal component effects at both the observation- and individual-level, from a BART-estimated conjoint model.
}
\details{
The OMCE estimates are the result of subtracting the predicted value of each observation under the reference-level category from the predicted value of each observation under the given attribute level.
If an attribute has \code{k} levels, then this will yield \code{k-1} estimates per observation.
The IMCE is the average of the OMCEs for each individual within the data.
}
\seealso{
\code{\link[=cjbart]{cjbart()}}
}
