% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.cjbart}
\alias{plot.cjbart}
\title{Plot Marginal Component Effects of a \code{cjbart} Object}
\usage{
\method{plot}{cjbart}(x, covar, type = "imce", plot_levels = NULL)
}
\arguments{
\item{x}{Object of class \code{cjbart}, the result of running \code{\link[=OMCE]{OMCE()}}}

\item{covar}{Character string detailing the covariate over which to analyse heterogeneous effects}

\item{type}{Character string, either "imce" or "omce", to plot individual- and observation-level effects respectively.}

\item{plot_levels}{Optional vector of conjoint attribute names to plot. If not supplied, all attributes within the conjoint model will be plotted.}
}
\value{
Plot of marginal component effects.
}
\description{
Plots either the observation-level marginal component effect (OMCE) or the individual-level marginal component effect (IMCE). By default, all attribute-levels in the model are plotted.
}
