% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.cjbart.vimp}
\alias{plot.cjbart.vimp}
\title{Plot Variable Importance Matrix for Heterogeneity Analysis}
\usage{
\method{plot}{cjbart.vimp}(x, covars = NULL, att_levels = NULL, ...)
}
\arguments{
\item{x}{Object of class \code{cjbart.vimp}, the result of running \code{\link[=het_vimp]{het_vimp()}}}

\item{covars}{Optional vector of covariate names to plot. By default, all included covariates are shown.}

\item{att_levels}{Optional vector of attribute-levels to plot. By default, all attribute-levels are shown.}

\item{...}{Additional arguments (not currently used)}
}
\value{
Plot of covariate importance scores
}
\description{
Plots a heatmap of variable importance, across predicted IMCEs. By default, all attribute-levels and covariates in the model are plotted.
}
