% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjbart.R
\name{RMCE}
\alias{RMCE}
\title{Inspect Round-Level Marginal Component Effect (RMCE)}
\usage{
RMCE(imces)
}
\arguments{
\item{imces}{An object of class "cjbart", the result of calling the \code{cjbart::IMCE()} function}
}
\value{
\code{IMCE} returns a data frame of RMCEs.
}
\description{
\code{RMCE} calculates the round-level marginal component effects from a cjbart model.
}
\details{
The RMCE estimates are the result of averaging the OMCEs within each round, for each subject in the experiment.
The RMCE is the intermediate causal quantity between OMCEs and IMCEs, and can be useful for inspecting whether there are any carryover or stability issues across rounds.
}
\seealso{
\code{\link[=cjbart]{cjbart()}} and \code{\link[=IMCE]{IMCE()}}
}
