% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.cjbart}
\alias{summary.cjbart}
\title{Summarizing \code{cjbart} Marginal Component Effect Estimates}
\usage{
\method{summary}{cjbart}(object, ...)
}
\arguments{
\item{object}{Object of class \code{cjbart}, the result of running \code{\link[=OMCE]{OMCE()}}}

\item{...}{Further arguments (not currently used)}
}
\value{
Tibble summarising the average marginal component effect, the minimum and maximum values, and standard deviations for each attribute-level.
}
\description{
\code{summary} method for class "cjbart"
}
