% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.cjbart}
\alias{plot.cjbart}
\title{Plot Marginal Component Effects of a \code{cjbart} Object}
\usage{
\method{plot}{cjbart}(x, plot_levels = NULL, se = TRUE, covar = NULL, ...)
}
\arguments{
\item{x}{Object of class \code{cjbart}, the result of running \code{\link[=OMCE]{OMCE()}}}

\item{plot_levels}{Optional vector of conjoint attribute names to plot. If not supplied, all attributes within the conjoint model will be plotted.}

\item{se}{Boolean determining whether to show an estimated 95\% confidence interval}

\item{covar}{Character string detailing the covariate over which to analyse heterogeneous effects}

\item{...}{Additional arguments for plotting the marginal component effects (see below).}
}
\value{
Plot of marginal component effects.
}
\description{
Plots observation-level or individual-level marginal component effects (OMCE and IMCE respectively). By default, all attribute-levels in the model are plotted.
}
