% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.cjbart}
\alias{summary.cjbart}
\title{Summarizing \code{cjbart} Marginal Component Effect Estimates}
\usage{
\method{summary}{cjbart}(object, ...)
}
\arguments{
\item{object}{Object of class \code{cjbart}, the result of running \code{\link[=IMCE]{IMCE()}}}

\item{...}{Further arguments (not currently used)}
}
\value{
Data frame summarizing the average marginal component effect (AMCE), the minimum and maximum values, and standard deviations for each attribute-level.
}
\description{
\code{summary} method for class "cjbart"
}
\note{
To calculate the AMCE with Bayesian credible intervals, please use the \code{AMCE()} function instead.
}
\seealso{
\code{\link[=AMCE]{AMCE()}}
}
